requirements.txt
faster-whisper
watchdog
pystray
Pillow
plyer
customtkinter
pyinstaller
